pyproject.toml
[tool.poetry]
name = "todo_api"
version = "0.1.0"
description = ""
authors = ["adriangube <adriangubedev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic-settings = "^2.7.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
email-validator = "^2.2.0"
alembic = "^1.14.0"
pytest = "^8.3.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "todo_api.main:start"
prod = "todo_api.main:start"

[tool.pylint]
disable = ["missing-function-docstring", "missing-module-docstring", "missing-class-docstring"]
